requirements.txt
flask
joblib
scikit-learn
pillow
numpy
